refactor: reduce custimization of hub project setup follow more standard layouts and usage of sea migrations. boot redpanda with docker compose.
