+++ .dockerignore
@@ -1,4 +1,4 @@
 /.git
 /.github
 /target
-.env.dev
+.env
