+++ .env
@@ -1,1 +1,3 @@
-DATABASE_URL=postgres://postgres:holaplex@localhost:5437/hub-orgs
+DATABASE_URL=postgres://postgres:holaplex@localhost:5437/orgs
+POSTGRES_DB=orgs
+POSTGRES_PASSWORD=holaplex
