--- .env.dev
@@ -1,3 +1,0 @@
-DATABASE_URL=postgres://postgres:holaplex@localhost:5437/hub-orgs
-DATABASE_NAME=hub-orgs
-DATABASE_PASSWORD=holaplex
