+++ docker-compose.yml
@@ -2,14 +2,43 @@ version: '3.8'
 services:
   db:
     image: postgres:15.1
-    container_name: hub_orgs_dev_db
-    environment:
-      - POSTGRES_DB=${DATABASE_NAME}
-      - POSTGRES_PASSWORD=${DATABASE_PASSWORD}
+    container_name: web
+    env_file:
+      - .env
     ports:
-      - '5437:5432'
+      - 5437:5432
     volumes: 
       - hub_orgs_db:/var/lib/postgresql/data
+  redpanda:
+    # NOTE: Use the latest version here!
+    image: docker.redpanda.com/vectorized/redpanda:latest
+    container_name: redpanda-1
+    command:
+    - redpanda
+    - start
+    - --smp
+    - '1'
+    - --reserve-memory
+    - 0M
+    - --overprovisioned
+    - --set redpanda.empty_seed_starts_cluster=false
+    - --set redpanda.auto_create_topics_enabled=true
+    - --seeds "redpanda-1:33145"
+    - --kafka-addr
+    - PLAINTEXT://0.0.0.0:29092,OUTSIDE://0.0.0.0:9092
+    - --advertise-kafka-addr
+    - PLAINTEXT://redpanda:29092,OUTSIDE://localhost:9092
+    - --pandaproxy-addr
+    - PLAINTEXT://0.0.0.0:28082,OUTSIDE://0.0.0.0:8082
+    - --advertise-pandaproxy-addr
+    - PLAINTEXT://redpanda:28082,OUTSIDE://localhost:8082
+    - --advertise-rpc-addr redpanda-1:33145
+    ports:
+    - 8081:8081
+    - 8082:8082
+    - 9092:9092
+    - 28082:28082
+    - 29092:29092
 volumes:
   hub_orgs_db:
     driver: local
