+++ migration/src/main.rs
@@ -2,5 +2,5 @@ use sea_orm_migration::prelude::*;
 
 #[async_std::main]
 async fn main() {
-    cli::run_cli(hub_orgs::migrations::Migrator).await;
+    cli::run_cli(migration::Migrator).await;
 }
